Create Actor, New AutocompleteComponent
implemented create actor:
Name,description and Movie are required fields. AutocompleteComponent used to search for movies so it can be selected in the create form

implemented autocomplete comp:
takes the typed chars and searches an object for results and suggest them to the user. 

querymessage are now shown onSucces.:
changed the Messages that is printed when updating/creating new items based on the action. e.g when updating prints "Successfully updated"